Add separate hosting environment for end-to-end testing
This hosting environment will be used to run the web application for
end-to-end testing.

It overrides the database connection string so that end-to-end tests can
be run locally in development without interfering with the default
application database.

It also specifies a fake SendGrid API so that emails cannot be
accidentally sent from end-to-end tests [1]. When we come to write
end-to-end tests for scenarios that do generate emails, we'll need the
option to run the app with a fake `IEmailSender` implementation instead.
However, that's not something we need to address in this branch.

[1] `EmailOptions.ApiKey` is marked as required so the application won't
run if no key is specified at all