Run end-to-end tests in CI
I moved the 'build assets' step so that the assets are built before the
end-to-end tests are run.

I also modified the `webServer` command in `playwright.config.ts` to
avoid rebuilding the application before launch (as the entire solution
is already built using the Release configuration earlier in the
workflow).

The 'Install Playwright browsers' step is relatively time consuming,
currently taking approximately 45s to complete on the GitHub-hosted
Ubuntu runner.

One option recommended in the Playwright docs for avoiding this time
consuming step is to run the tests in a containerized job [1]. This
could well be the best option for us in the longrun, but I've discounted
it for now as it would involving preparing our own Docker image with all
of the dependencies needed to run the app (including .NET and MySQL
server).

Another option I experimented with with was running the end-to-end tests
in a separate parallel job. However, since most of the running time for
the existing CI workflow is spent downloading dependencies (including
downloading NuGet packages as part of the .NET build step) which would
need to be done in both jobs, the time saved by this parallelisation is
essentially limited to the running time of the unit tests and linters,
which is currently about 40 seconds (or 20% of the total running time).

At the moment this doesn't like enough of a saving to justify the
additional overhead, but we should reconsider this option if the running
time of the unit tests and linters increases in the future.

Finally the configuration generated by the Playwright installer limits
the number of workers to 1 in CI (as recommended in the Playwright docs
to prioritise stability and reproducibility over speed [2]). As it
happens the GitHub-hosted Ubuntu runners only have two logical CPU cores
[3], and Playwright only uses half the number of logical CPU cores by
default [4], so this limit currently has no effect in practice.

[1] https://playwright.dev/docs/ci#docker
[2] https://playwright.dev/docs/ci#workers
[3] https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources
[4] https://playwright.dev/docs/api/class-testconfig#test-config-workers