Create and seed database when starting app for end-to-end testing
Creating the database on start up is strongly discouraged [1] in
production for a number of reasons, including the lack of opportunity to
inspect the changes before they are applied, and the risk that more than
one instance of the application may attempt to run the migrations at the
same time.

These concerns are much less relevant to the E2E environment where the
database can be considered ephemeral, and no more than one instance of
the application is expected be running at a time.

Because this code only creates and seeds the database if it doesn't
already exist, even if the hosting environment is incorrectly set to
'E2E', there is no risk of this code causing data loss or inserting seed
data into a production database.

[1] https://learn.microsoft.com/en-us/ef/core/managing-schemas/migrations/applying?tabs=dotnet-core-cli#apply-migrations-at-runtime