Move Playwright test results into a `.playwright` directory
Because we have multiple types of tests in this project, Playwright's
default output directory [1], `test-results`, felt a little too generic.

Moving the output into a `.playwright` directory makes it obvious that
the files are associated with the Playwright end-to-end tests and
removes a little bit of clutter from the root of the web project.

[1] https://playwright.dev/docs/api/class-testconfig#test-config-output-dir