Add mutation for creating a recipe
I spent quite a bit of time deciding whether the mutation resolver or
the `IRecipeManager.AddRecipe` method should be responsible for
validating the recipe attributes.

Validating the attributes in the recipe manager would have the obvious
advantage that it would ensure that the validation rules are applied
consistently regardless of whether the recipe is being created via the
web UI or the API.

On the flip side, for recipes being created through the UI, the
validation attributes are already being enforced by ASP.NET MVC during
model binding. Running the validation checks in the recipe manager would
therefore introduce redundant processing.

Furthermore, relying on the recipe manager to validate the attributes
for recipes created via the API would likely involve throwing and
catching an exception on when there are validation errors which would
incur a performance penalty.

I therefore ultimately settled on calling the validator directly in the
resolver.