Enable mutation conventions
It is considered best practice for mutations to take a single input
object-typed argument called `input` and return a payload object that
exposes the record or records changed by the mutation and any domain
errors raised.

Although the only mutation we've added so far already conforms to these
rules, enabling the mutation conventions option in Hot Chocolate will
help us remain compliant as we add further mutations and minimize the
amount of boilerplate code needed to do so.