Add generic validator service
This validator is intended to provide similar functionality to the
`Validator` class in `System.Component.DataAnnotations` with a few key
improvements:

 - It exposes validation errors using the `ValidationErrors` type added
   in the previous commit which, amongst other things, provides access
   to the validation attribute that triggered each error
 - It will eventually support `IStringLocalizer`-based localization of
   messages
 - It is non-static and implements an interface, making it suitable for
   dependency injection

Conversely it does not support some features that I don't expect to be
useful for this project including:

 - Breaking on the first error
 - Validation through the `IValidatableObject` interface
 - Skipping over validation attributes inherited from property types [1]

Like the `Validator` class in `System.Component.DataAnnotations`, this
validator does not recursively validate properties. That's because it's
easier and more efficient for the caller to explicitly validate nested
objects as required.

For efficiency, the type's properties and validation attributes are read
and cached on initialization. The validator therefore registered with
the service collection as a singleton so that the cached metadata for
each type is shared across all consumers.

[1] https://github.com/dotnet/runtime/blob/5535e31a712343a63f5d7d796cd874e563e5ac14/src/libraries/System.ComponentModel.Annotations/src/System/ComponentModel/DataAnnotations/ValidationAttributeStore.cs#L235