Localize validation errors
`ValidationAttribute` and the `System.ComponentModel.DataAnnotations`
APIs in general, do not support localizing strings with
`IStringLocalizer`. Indeed that's one of the reasons for adding our own
`ILocalizer<T>` service.

The approach I've taken here is based on how ASP.NET Core MVC provides
localization support for data annotations [1].

If the `ErrorMessage` property is set on the validation attribute then
this is assumed to be the resource name that should be used to retrieve
the translated string from the `IStringLocalizer<T>` for the validated
object type.

The switch statement in `GetMessageArguments` is then used to get the
message arguments based on the validation attribute type (the
`DataAnnotationsModelValidator` in ASP.NET MVC does essentially the same
thing through adapter classes instead [2]). For simplicity only
validation attributes used in the application are included in the switch
statement.

[1] https://github.com/dotnet/aspnetcore/blob/3f1acb59718cadf111a0a796681e3d3509bb3381/src/Mvc/Mvc.DataAnnotations/src/DataAnnotationsModelValidator.cs#L87
[2] https://github.com/dotnet/aspnetcore/blob/3f1acb59718cadf111a0a796681e3d3509bb3381/src/Mvc/Mvc.DataAnnotations/src/DataAnnotationsModelValidator.cs#L134