Add GraphQL type to represent a validation error
Although the GraphQL type is named `ValidationError`, as this class will
be used alongside `Buttercup.Application.Validation.ValidationError`,
I've named the runtime type `InputObjectValidationError`.

The values in the `ValidationErrorCode` enum correspond to the
validation attributes currently in use in the `RecipeAttributes` class.
Additional codes can be added as and when they are needed.