Use value equality when comparing objects in API responses
`JsonAssert.ValueEquals` was intended to be used to check that a JSON
element is equal to an expected value once deserialized. Because of the
deserialization step the expected and actual values are guaranteed to be
distinct instances and so the assertion has to be based on value
equality.

Despite this requirement, the method was actually using `Assert.Equal`
for the assertion when, except for collections, essentially uses the
default equality check for the type. For most object types, this means
reference equality.

This went unnoticed up to now because the method was only being used
with anonymous classes and record types, both of which use value
equality by default.

However, we're about to use it to compare validation errors returned by
a mutation, and these include an array valued property that would be
compared using reference equality.

This commit therefore replaces the `ValueEquals` method with on that
uses `Assert.Equivalent` for the comparison instead.