Add fixture to facilitate use of GraphQL API in end-to-end tests
We're about to start using the GraphQL API to setup and cleanup test
data in end-to-end tests.

Currently all this fixture does is create a URQL client [1] with an auth
exchange that takes care of obtaining and injecting an access token into
requests.

However, instead of consuming this client directly, my intention is to
add a function to the fixture each time a new GraphQL operation is
needed in the tests.

[1] I've chosen to use URQL as it appears to be the most popular and
well maintained lightweight GraphQL library currently available for
TypeScript.