Bump stylelint to 16.2.1
stylelint 15.10.0 introduced [1] a new `media-query-no-invalid` rule
that raises false positives when LESS variables are used in media
queries. Indeed the documentation for the rule states that it should be
turned off for SASS and LESS, so that's what I've done.

It was also necessary to upgrade stylelint-config-standard 36.0.0 to
satisfy its peer dependency on stylelint.

[1] https://github.com/stylelint/stylelint/pull/6963