Redefine unique index on user email
Previously the unique constraint on user email was expressed as an
alternate key in EF Core.

However, in EF Core, alternate keys provide additional semantics over
unique indexes (they are read-only and can be used as the target of
foreign key) that are not applicable in this case.

Defining the unique constraint as a unique index instead has no impact
on the MySQL schema other than name of the underlying unique key.

Before:

```
CREATE TABLE `users` (
  ...
  UNIQUE KEY `ak_users_email` (`email`)
)
```

After:

```
CREATE TABLE `users` (
  ...
  UNIQUE KEY `ix_users_email` (`email`)
)
```