Add method for finding a non-deleted recipe
In the next few commits I'll be updating various controller actions to
use this method to ensure that various controller actions only retrieve
and display non-deleted recipes.

Instead of throwing exceptions, I've chosen to return null when the
recipe does not exist or is soft-deleted. That's partly because throwing
exceptions is relatively costly, but also because it'll make it easier
to handle both cases together in the controller layer.