Remove `HandleNotFoundExceptionAttribute`
The POST action for editing a recipe was the last remaining controller
action that relied on this attribute filter to convert
`NotFoundException` to an HTTP 404 not found result.

Furthermore that controller action was already explicitly handling
`ConcurrencyException` and `SoftDeletedException`, so it was arguably
preferable to explicitly handle `NotFoundException` in the controller
action even if `HandleNotFoundExceptionAttribute` was still useful
elsewhere.