Use built-in concurrency checking mechanism for recipes
Previously `UpdateRecipe` used EF Core's `ExecuteUpdateAsync` method
combined with where clauses to enforce optimistic concurrency control
and prevent the modification of soft-deleted recipes.

This commit refactors the method to use EF Core's `SaveChangesAsync`
method and built-in optimistic concurrency control mechanism [1]
instead.

Because EF Core's `SaveChangesAsync` method relies on change tracking,
we now have to retrieve the current version of the recipe upfront. EF
Core's built-in concurrency control mechanism also only allows us to
check that certain columns contain the corresponding original value in
the change tracker. This means that we now have to explicitly check that
`Recipe.Deleted` is null and that `Recipe.Revision` matches the supplied
base revision before calling `SaveChangesAsync` [2].

The `ConcurrencyCheck` attributes on the `Deleted` and `Revision`
properties ensure that the update is not applied if the value of either
property has changed since the retrieve (which could happen in the case
of a near-simultaneous update or delete). If this occurs EF Core will
throw `DbUpdateConcurrencyException`, which we handle by rerunning the
method, causing the checks to be repeated against a freshly
retrieved version of the record.

It's worth noting that in the majority of cases this new implementation
results in an extra database query, but this feels like a reasonable
trade-off for simpler Entity Framework code.

Scenario                | Queries before  | Queries after
----------------------- | --------------- | ------------------------
Successful update       | UPDATE          | SELECT > UPDATE
Stale/deleted (no race) | UPDATE > SELECT | SELECT
Stale/deleted (race)    | UPDATE > SELECT | SELECT > UPDATE > SELECT

[1] https://learn.microsoft.com/en-us/ef/core/saving/concurrency?tabs=data-annotations#optimistic-concurrency
[2] Alternatively we could explicitly override the original values for
those properties in the change tracker to null and the supplied base
revision, respectively. However we would still eventually need to
explicitly check those values to determine which exception to throw in
the case of an concurrency violation.