Add recipe revision model and schema
Unfortunately the new `Revisions` property effectively breaks the value
equality of the `Recipe` record type. Because every instance of `Recipe`
is initialised with `Revisions` set a new instance of `List<T>`, and
because `List<T>` is a reference type, the `Revisions` property for two
distinct instances of `Recipe` are never considered equal, even if they
contain the same records.

I therefore had to update unit tests that previously relied on the value
equality of the `Recipe` type to use `Assert.Equivalent` instead of
`Assert.Equals` when comparing `Recipe` instances.

For collections I had to use `Assert.Collection` instead as
`Assert.Equivalent` ignores the ordering of elements in collections.

I've defined `Recipe.Revisions` as a read-write property because Hot
Chocolate projections currently need a public setter on fields they
operate on [1]. If the property was read-only the corresponding field in
the GraphQL API would always return an empty collection.

[1] https://github.com/ChilliCream/graphql-platform/issues/3486