Add dedicated query method for recipe 'show' view
Adding comments to the recipe 'show' view will involve at lot more eager
loading of related records. This tips the balance in favour of having a
completely separate method for finding and loading the recipe and
related data for the view (rather than adding further parameters to a
single `FindRecipe` method).