Organise view models into sub-namespaces by controller
This means that view models are now organised in the directory structure
as the corresponding views.

I chose to suppress CA1716 (identifiers should not match keywords) for
`Buttercup.Web.Models.Error` because I felt this violation was justified
in the interests of keeping the namespace name consistent with the
corresponding controller name.