Replace object type extensions with code-first object types
At the moment we're only using these type extension to ignore a handful
of properties, which is arguably more concisely done by configuring the
object type descriptor directly.

In a few commits time, however, we'll also start attaching authorisation
policies to existing properties, which is definitely easier to do with
code-first object types.