Add new identity to principal in `authenticate` mutation
We're about to add authorisation policies to certain user fields
including `email`. To avoid authorisation errors when those fields are
selected in the `authenticate` query, we need to ensure that the
authenticated user's claims are included in the current principal.

I'm not aware of any way to replace the current principal or its
identities from within a resolver. I believe the only option is to add a
new identity to the current principal instead.

This means that, if the authenticate mutation is invoked with a valid
bearer token, the principal will end up containing two identities. This
won't be a problem so long as the authorization policies are written to
check for the presence of a specific claim as opposed to checking the
value of the first claim of a specific type.