Restrict access to sensitive user fields
For security and privacy reasons, the `email`, `passwordCreated` and
`isAdmin` fields on the user type should only be accessible to admins
and users querying their own user record.

I've marked the `User.Id` field as always projected so that it can be
accessed in the authorization handler even if it isn't selected in the
query.