Add support for markdown in comments
The main reason for supporting markdown is that text separated by blank
lines will now intuitively be rendered as paragraphs (previously all
whitespace was collapsed).

As a bonus, it also allows us to safely support basic formatting options
and hyperlinks for advanced users without all the complexity of a rich
text editor.