Select user `name` field instead of `email` in API tests
The user `email` field now has an authorisation policy [1] to ensure
that non-admins are not able to see other users' email addresses.

In theory this should have caused many of the API tests that involved
the `email` field to start failing. The only reason that didn't happen
is that, just by luck, the internal counter in the model factory
happened to always be set to an even number when `BuildUser` was setting
`IsAdmin` in the affected tests.

That's obviously not something we'd like to continue relying on, so this
commit updates the tests to select the `name` property instead of
`email`.

[1] de0b5c3057213f410de528a2e953455b2eae795a