Allow end-to-end tests to make API calls on behalf of different users
Previously all API calls made through the `api` fixture were
authenticated using the credentials for the `e2e-admin` user.

However, in the next commit, I want to add an end-to-end test that use
the API to create a comment as a non-admin user.

This commit therefore changes the `api` fixture to take the form of a
function that takes a username as an argument. This means that a single
test can now invoke the fixture multiple times to make API calls on
behalf of different users.