Move all end-to-end Xunit tests into a single collection
Previously each class that extended `EndToEndTests` had `AppFactory` as
a class collection, which meant each test class had its own copy of the
database.

Moving all of these tests into a collection and making `AppFactory` to a
collection fixture instead means that we now create only one copy of the
database for all of the end-to-end tests.

This comes at the cost of no longer being able to run the end-to-end
tests in parallel with each other. However, since these tests only make
up a small proportion of the tests in the solution, that's unlikely to
negatively impact the test runner's ability to make best use of all
available CPU threads.

On my desktop PC (Ryzen 5 3600 with 12 threads), this change has reduced
the average running time for all tests in the solution by only 0.5
seconds (7.6%). However, on my laptop (Core i5 6300U with 4 threads) the
average running time is reduced by a more noticeable 4 seconds (17.5%).