Add Redis connection factory and manager
`IRedisConnectionManager` is a singleton service responsible for
initialising the connection to the Redis cluster and recreating it in
the rare cases [1] where it fails to reconnect after a transient
connection failure.

It is based heavily on the `RedisConnection` class in the Azure Cache
for Redis sample for ASP.NET Core [2] with a few notable differences:

- `IRedisConnectionManager` is not responsible for providing any retry
  logic. Instead, consumers are expected to call
  `IRedisConnectionManager.CheckException` when an exception occurs
  making a call to Redis. This return value indicates whether the
  exception could have been caused by a transient connection error,
  which the caller can use to decide whether to perform a retry.
- I've added separate configuration options
  (`DroppedConnectionGracePeriod` and `DroppedConnectionEpisodeTimeout`)
  for the two conditions checked against `RedisErrorThreshold` in the
  Azure Cache for Redis sample
- `RedisConnectionManager` implements `IAsyncDisposable` instead of
  `IDisposable`

I've abstracted the creation of Redis connections into a separate
`IRedisConnectionFactory` service purely to facilitate unit testing of
`RedisConnectionManager`. This interface is not intended to be consumed
directly and is therefore internal to the assembly.

[1] https://github.com/StackExchange/StackExchange.Redis/issues/559
[2] https://github.com/Azure-Samples/azure-cache-redis-samples/blob/main/quickstart/aspnet-core/ContosoTeamStats/RedisConnection.cs