Exclude soft-deleted comments from `recipe.comments`
If a user deletes one of their comments, it would seem reasonable for
them to expect that it is no longer visible to other users (apart from
maybe administrators).

I considered a number of options for achieving this including:

1) Only including delete comments in this list if the current user is an
   administrator
2) Continuing to include all comments in the list, but with the comment
   body redacted on soft-deleted comments if the current user is not an
   administrator
3) Excluding deleted comments from the list completely (and later adding
   a separate `deletedComments` field for admins)

I rejected option 1 because I wasn't keen on the field having different
semantics depending on the user's permissions.

Option 2 had the benefit that it would allow the client to indicate
where in a thread comments had been deleted, and I did have some success
with a proof of concept implementation using custom middleware. However
I didn't think the extra complexity was justified at this point. I was
also mindful that we're likely to eventually enable paging on this
field, at which point including deleted comments in the list would
result in inconsistent page sizes when deleted comments are not rendered
in the UI.

Option 3 therefore felt like the simplest and safest option.

One important thing to note is that this change does introduce N+1
queries when querying comments across multiple recipes. However I don't
think that's likely to be a frequent use case in practice, and we'd be
unable to avoid this tradeoff if we were to add paging later anyway.