Prevent non-admins for retrieving soft-deleted recipes by ID
In an earlier branch we updated the `recipes` field to not return
soft-deleted recipes [1] and added a separate `deletedRecipes` field [2]
to allow soft-deleted recipes to be queries by admins only.

However, until this commit, it was still possible for non-admins to find
and retrieve deleted recipes by enumerating on the `recipe` field.

[1] a2d0407081ccd04a99dd867c8126dc3c9261efb9
[2] ce000335c5ff064d888175ad38c1f6b5f301acc6