Add extension method to 'unwrap' resource in authorization handlers
When authorizing access to a field in the GraphQL API,
`AuthorizationHandlerContext.Resource` is an `IMiddlewareContext`.
Provided that `ApplyPolicy.AfterResolver` is used, this exposes the
field value (resolver result) via its `Result` property.

This method – which returns `IMiddlewareContext.Result` when
`AuthorizationHandlerContext.Resource` is an `IMiddlewareContext` or
`AuthorizationHandlerContext.Resource` otherwise – will therefore make
it easier to implement authorization handlers for policies that are both
attached to GraphQL fields and evaluated directly against the resource
elsewhere.