Rename and specialize 'SelfOrAdmin' authorization policy
One thing that troubled me with the old 'SelfOrAdmin' authorization
policy and requirement was that when it was applied to a GraphQL field,
the subject was assumed to be the parent resolver result, but in all
other contexts the subject was assumed to be the current resource.

This was inconsistent with the 'AdminWhenDeleted' policy and requirement
which, when applied to a GraphQL field, assumes the subject to be the
current resolver result.

Since the policy has so far only been used on user fields in the GraphQL
API, I've decide to resolve this inconsistency by restricting its usage
to that specific case and renaming it accordingly.