Ensure that CI workflow uses latest SDK in 8.0.1xx feature band
Previously, in the absence of a `global.json` file to specify the
desired SDK version, the highest installed SDK version was being used.

On the `ubuntu-22.04` hosted runners, this is currently v8.0.401.

Meanwhile, the Canonical .NET SDK packages for Ubuntu are pinned to the
1xx feature band for the lifetime of each major .NET version [1], which
means that the highest installed SDK version on my local dev machine is
v8.0.108.

The `global.json` now explicitly specifies the SDK version '8.0.100'.
The default 'roll-forward' policy when an SDK version is specified is
'latestPatch' [2], so this ensures that the highest installed patch
version in the 8.0.1xx feature band is used in all environments.

This is important as there can be breaking changes between feature bands
(as demonstrated by the issue addressed by the previous commit).

[1] https://github.com/dotnet/core/discussions/9258
[2] https://learn.microsoft.com/en-us/dotnet/core/tools/global-json#matching-rules