Reimplement admin-only sort field authorization
`RoleRestrictedOrderByFieldsRequirement` that the previous
implementation was based on had a couple of notable defects and
limitations.

The most serious of these was that the authorization handler didn't take
into account GraphQL's input coercion rule [1] for lists. This resulted
in the authorization checks being completely bypassed if the order
argument was provided in the form of a single object literal, such as in
the following example:

    query {
      users(order: { email: ASC }) {
        id, name
      }
    }

The authorization handler was also unable to verify nested fields, which
effectively prevented us from allowing recipes and comments to be sorted
by author, for example.

This commit seeks to address these issues by replacing
`RoleRestrictedOrderByFieldsRequirement` with a new
`AdminOnlySortFieldsRequirement` that properly recurses through lists
and object literals and uses a new '@adminOnly' directive to identify
restricted sort fields as part of the GraphQL schema.

The new `AdminOnlySortFieldsRequirement` is also hardcoded to require
the admin role, which felt like a sensible simplification to make given
that the application doesn't have any other roles at present.

The tests for the new authorization handler are more like integration
tests that execute dummy queries through the Hot Chocolate request
executor. This is because the handler references
`IMiddlewareContext.Selection.Arguments` which cannot be mocked or
initialized externally with arguments. As it happens, I think this
probably makes the tests more valuable anyway, especially given how
strongly tied the new logic is to the intricacies of the GraphQL type
system.

[1] https://spec.graphql.org/October2021/#sel-HAHXPBAABABOnmF