Allow sorting on most recipe properties
I originally [1] opted to be restrict sorting to a smaller number of
indexed properties. However, in practice, the table is unlikely to grow
large enough for sorting on other fields to pose any real performance
issues.

As a result of this change it is now possible to sort recipes by user
properties on `createdByUser`, `modifiedByUser` and `deleteByUser`. I
therefore had to add the `AdminOnlySortFields` authorisation policy to
the `recipes` field to prevent sorting on admin-only fields. The
`deletedRecipes` field, on the other hand, is itself an admin-only
field, so didn't need updating.

[1] b276b423fa65feca7261b2589912e5c92c6b4078