Use raw string literals for multiline GraphQL queries
The main benefit of using raw string literals (""") instead of verbatim
literals (@") is that the former does not require us to escape double
quotes. A secondary benefit is that any leading whitespace before the
closing quotes is removed from every line, which preserves the desired
indentation in the resulting string value.