Add middleware to guarantee unambiguous sort order
The purpose of this middleware is to ensure that the field always has an
unambiguous sort order, which is particularly import for pagination [1].
It does this by appending `id ASC` as a final tie-breaker in the order
by clause.

Unfortunately EF Core doesn't provide a simple way to append a sort
column to an `IQueryable<T>` that may or may not already have ordering
applied. We therefore have to use an expression visitor look for any
calls to `OrderBy` or `OrderByDescending` and then call either `OrderBy`
or `ThenBy` accordingly .

[1] When pagination is used without explicit ordering, EF Core logs the
following warning:

```
warn: Microsoft.EntityFrameworkCore.Query[10102]
      The query uses a row limiting operator ('Skip'/'Take') without an
      'OrderBy' operator. This may lead to unpredictable results. If the
      'Distinct' operator is used after 'OrderBy', then make sure to use
      the 'OrderBy' operator after 'Distinct' as the ordering would
      otherwise get erased.
```