Do not generate GraphQL documentation from XML documentation
By default Hot Chocolate automatically generates [1] GraphQL API
documentation using the XML documentation tags on the underling C#
types.

However the results of this are not ideal as the XML documentation for
the C# properties typically takes the form "Gets or sets the..." whereas
the corresponding GraphQL fields that inherit the documentation are
specifically used for querying, sorting or filtering.

[1] https://chillicream.com/docs/hotchocolate/v13/defining-a-schema/documentation/#xml-documentation