Refactor `UseTieBreakSortByIdAttribute` to use `OnAfterSortingApplied`
HotChocolate 14 introduces [1] a couple of new methods on
`ISortingContext` that can be used to ensure a stable sort order on top
of any user-defined sort order.

Updating `UseTieBreakSortByIdAttribute` to use this new functionality
allows us to eliminate our custom expression visitor that was risked
getting out of sync with Hot Chocolate's internals.

[1] https://chillicream.com/blog/2024/08/30/hot-chocolate-14/#pagination