Return `PasswordAuthenticationResult` from `Authenticate` method
Up to now the UI and `Authenticate` GraphQL mutation have been able to
treat all authentication failures in the same way. However, when we add
rate limiting to password authentication attempts, both the UI and API
clients will need to be able to differentiate failures caused by the
rate limit being exceeded from those caused by invalid credentials.

To pave the way for this, this commit updates the
`IPasswordAuthenticationService.Authenticate` method to return a
`PasswordAuthenticationResult` object that contains the authenticated
user on success, or an enum value indicating the cause of failure.