Migrate to xUnit 3
The changes in this commit are based on the steps described in the
official migration guide [1] and by comparing the test project files to
the `xunit3` template [2].

The migration process normally involves adding a `content` tag for
xunit.runner.json to the project file. I omitted this from
'Buttercup.Web.Tests.csproj', however, because that project uses the
'Microsoft.NET.Sdk.Web' SDK which includes files from the project
directory as content automatically. Indeed, attempting to add the
`content` tag to that project resulted in the following build error:

    error NETSDK1022: Duplicate 'Content' items were included. The .NET
    SDK includes 'Content' items from your project directory by default.
    You can either remove these items from your project file, or set the
    'EnableDefaultContentItems' property to 'false' if you want to
    explicitly include them in your project file. For more information,
    see https://aka.ms/sdkimplicititems. The duplicate items were:
    'xunit.runner.json'

The most notable changes required to the tests were:

1) Passing through the cancellation token to asynchronous methods that
   take one, which is necessary to avoid following compiler warning:

    warning xUnit1051: Calls to methods which accept CancellationToken
    should use TestContext.Current.CancellationToken to allow test
    cancellation to be more responsive.
    (https://xunit.net/xunit.analyzers/rules/xUnit1051)

2) Updating test classes and fixtures that implement `IAsyncLifetime` to
   return a `ValueTask` instead of a `Task` from the `InitializeAsync`
   and `DisposeAsync` methods (which is necessary as a result of
   `IAsyncLifetime` now extending `IAsyncDisposable`).

[1] https://xunit.net/docs/getting-started/v3/migration
[2] https://github.com/xunit/xunit/blob/main/src/xunit.v3.templates/templates/project/xunit3/cs/TestProject.csproj