Add config setting to explicitly enable telemetry
The Application Insights package we're currently using does nothing when
no connection string is provided, whereas the newer OpenTelemetry
package we're about to migrate to raises an error on start up in the
same situation.

This commit therefore introduces an `EnableTelemetry` configuration
setting that controls whether the telemetry services are added to the
application on start up. This is set to true in the base configuration
(so that the services are added in production) and then overridden to
false for the development and E2E hosting environments where we don't
generally want to enable telemetry.