Use more specific key for 'too many authentication attempts' message
This error message includes advice to reset your password, which only
makes sense when it's the authentication rate limit that's been
exceeded.

In the next commit I'll be adding rate limiting to password reset
requests too and so, although there is currently no mutation for
password reset requests, it now feels more appropriate to give this
message a resource key that explicitly refers to authentication
attempts.