Replace `LogAssert.HasEntry` with `LogAssert.SingleEntry`
In single parameter overload of the new `SingleEntry` method is intended
to be used in cases where exactly one message is expected to be logged.
Unlike the old `HasEntry` method, this overload ensures that the test
fails if other messages are unexpectedly logged.

The two-parameter overload is intended to be used in cases where
multiple log messages are expected, but only one with a specific event
ID.

Both overloads return a `FakeLogRecordAssert` object that provides a
fluent API for verifying properties on the matched log entry. This
provides greater flexibility over which properties are checked, produces
more helpful failure messages, and helps improve the readability of the
tests (albeit at the expense of a little more verbosity).