Simplify event ID allocation
Previously event IDs were typically allocated within a unique range per
component. However, the allocation of the ranges always felt a little
arbitrary and unwieldly.

As it happens, all that really matters is that each event ID is unique
within the scope of the log category, which is almost always the same as
the class name.

We can therefore simplify the allocation of event IDs by starting the
numbering at 1 within each class.