Send emails using Azure instead of SendGrid
SendGrid is retiring its free plans [1] so I'm switching to Azure
Communication Services instead.

I had to change the from address to 'donotreply@buttercup.doubliez.net'
because using anything other than 'DoNotReply' as the from address in
Azure Communication Services requires a quota increase that I'm not
confident to be able to obtain for a private hobbyist project like this
one.

[1] https://www.twilio.com/en-us/changelog/sendgrid-free-plan