Allow Azure SDK to authenticate using managed identities
The recommended approach for authenticating access to Azure resources
from an Azure-hosted app is to use managed identities [1].

Since the Buttercup app is currently hosted on an Azure VM that hosts
other unrelated applications, in the interests of being able to keep the
permissions for each app separate, I've chosen to create a user-assigned
managed identity specifically for the permissions needed by the
Buttercup app.

This commit therefore inserts a `ManagedIdentityCredential` into the
Azure client credential chain. This can be configured to use a
user-assigned managed identity by specifying the corresponding client ID
in the `Azure:ManagedIdentity:ClientId`` configuration setting, or the
system-assigned managed identity by leaving the same configuration
setting blank.

I've placed `ManagedIdentityCredential` after `EnvironmentCredential` in
the chain as the latter is able to fail fast in the absence of relevant
environment variables.

[1] https://learn.microsoft.com/en-us/dotnet/azure/sdk/authentication/#authentication-for-azure-hosted-apps